Add files via upload
Predicting how the stock market will perform is one of the most difficult tasks to do. There are so many factors involved in the prediction; physical vs. physiological, rational and irrational behavior, etc. All these aspects combined makes share prices volatile and very difficult to predict, with a high degree of accuracy. 
    Broadly, stock market analysis is divided into two parts – Fundamental and Technical Analysis.

Fundamental Analysis involves analyzing the company’s future profitability on the basis of its current business environment and financial performance
visualizations, as well as prediction.

Technical Analysis, on the other hand, includes reading the charts and using statistical figures to identify the trends in the stock market.
Problem Statement

Mr. Chinedu wants to invest in XYZ stock, but he is very skeptical due to the volatility in stock prices and because he has had losses in the past. Hence, He is requesting Coven Data company to help him build a model that will predict the future price of XYZ company shares.
DEFINITION OF TERMS
To demonstrate a basic method, we will start with just a few simulations. First LET US define the variables we'll be using in the XYZ Stock DataFrame.
 OPENING PRICE: 
    	  This is the starting price at which the stock is traded on a particular day.

 HIGH PRICE:  
  	 This refers to the highest price at which a stock is been traded in a day, which is usually  higher than  the closing price.

 LOW PRICE:    
 	This refers to the lowest price at which a stock is been traded on a particular business day.

 CLOSING PRICE:  
 	This is the last/final price at which a stock trades in a day.

 VOLUME PRICE:
 	This is total number of shares bought or sold  in  particular day i.e. the Turnover  which is  varies with company on a given date.